+++ extensions/camel-k/client/src/main/java/io/fabric8/camelk/client/V1alpha1APIGroupClient.java
@@ -1,0 +1,42 @@
+/**
+ * Copyright (C) 2015 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.fabric8.camelk.client;
+
+import io.fabric8.kubernetes.client.BaseClient;
+import io.fabric8.kubernetes.client.Config;
+import io.fabric8.kubernetes.client.dsl.MixedOperation;
+import io.fabric8.kubernetes.client.dsl.Resource;
+import io.fabric8.camelk.client.dsl.V1alpha1APIGroupDSL;
+import io.fabric8.camelk.client.internal.v1alpha1.KameletOperationsImpl;
+import io.fabric8.camelk.v1alpha1.Kamelet;
+import io.fabric8.camelk.v1alpha1.KameletList;
+import okhttp3.OkHttpClient;
+
+public class V1alpha1APIGroupClient extends BaseClient implements V1alpha1APIGroupDSL {
+
+  public V1alpha1APIGroupClient() {
+    super();
+  }
+
+  public V1alpha1APIGroupClient(OkHttpClient httpClient, final Config config) {
+    super(httpClient, config);
+  }
+
+  @Override
+  public MixedOperation<Kamelet, KameletList, Resource<Kamelet>> kamelets() {
+    return new KameletOperationsImpl(httpClient, getConfiguration());
+  }
+}